requirements.txt
requests
duckduckgo-search
beautifulsoup4
python-dotenv
openai
google-generativeai
anthropic
jinja2
schedule
tavily-python
